test: move the catalog-fixture half of the author pin to scripts/__tests__
packages/plugin-timeline's test tsconfig carries no node types, so the
fs-reading assertions belonged where the repo's other catalog test lives.
The registry and designer assertions stay in the package.
